Prevent dark mode from chaning on route - but there is a brief flicker in the nav bar toggle on route change
