+++ src/components/contact-form.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 import { useState } from 'react';
-import '../styles/global.scss';
 import { useForm } from 'react-hook-form';
 import axios from 'axios';
 import TextareaAutosize from 'react-textarea-autosize';
